+++ src/de/mhlz/MIPSInterpreter/Operations/IType/LWOperation.java
@@ -7,7 +7,7 @@ import de.mhlz.MIPSInterpreter.Operations.IOperation;
  * Created by mischa on 13.02.14.
  */
 public class LWOperation extends IOperation {
-	public LWOperation(Interpreter i, int rs, int rt, short immediate) {
+	public LWOperation(Interpreter i, int rs, int rt, int immediate) {
 		super(i, rs, rt, immediate);
 	}
