+++ src/de/mhlz/MIPSInterpreter/Operations/IType/SubIOperation.java
@@ -7,7 +7,7 @@ import de.mhlz.MIPSInterpreter.Operations.IOperation;
  * Created by mischa on 11.02.14.
  */
 public class SubIOperation extends IOperation {
-	public SubIOperation(Interpreter i, int rs, int rt, short immediate) {
+	public SubIOperation(Interpreter i, int rs, int rt, int immediate) {
 		super(i, rs, rt, immediate);
 	}
